chore: split expectedPackages publications to own file
