+++ meteor/server/publications/_publications.ts
@@ -1,6 +1,7 @@
 import './lib'
 
 import './buckets'
+import './expectedPackages'
 import './organization'
 import './peripheralDevice'
 import './rundown'
